MenuItem: create component
